+++ R/themes-project.R
@@ -6,22 +6,44 @@ gitignore_lines <- c(".Rproj.user", ".Rhistory", ".RData", ".DS_Store")
 #' Creates a theme that scaffolds a simple R analysis project with `main.R`,
 #' `README.md`, and `.gitignore`.
 #'
+#' @param data_dirs Logical. If `TRUE` (default), creates directories
+#'   `./data_raw` and `./data_processed`.
 #' @return A `prefab_theme` object.
 #' @export
 #'
 #' @examples
 #' r_analysis()
-r_analysis <- function() {
+r_analysis <- function(data_dirs = TRUE) {
   from_prefab <- from_package("prefab")
+  step_data_raw <- if (data_dirs) {
+    step_run(
+      fs::dir_create,
+      "data_raw",
+      .label = "fs::dir_create('data_raw')"
+    )
+  }
+  step_data_processed <- if (data_dirs) {
+    step_run(
+      fs::dir_create,
+      "data_processed",
+      .label = "fs::dir_create('data_processed')"
+    )
+  }
+  step_main <- from_prefab("r_analysis/main.R", "main.R", strategy = "skip")
+  step_readme <- from_prefab(
+    "r_analysis/README.md",
+    "README.md",
+    strategy = "skip",
+    data = "auto"
+  )
+  step_gitignore <- step_text(gitignore_lines, ".gitignore", strategy = "union")
+
   new_theme(
-    from_prefab("r_analysis/main.R", "main.R", strategy = "skip"),
-    from_prefab(
-      "r_analysis/README.md",
-      "README.md",
-      strategy = "skip",
-      data = list()
-    ),
-    step_text(gitignore_lines, ".gitignore", strategy = "union")
+    step_data_raw,
+    step_data_processed,
+    step_main,
+    step_readme,
+    step_gitignore
   )
 }
 
@@ -44,7 +66,7 @@ r_targets <- function() {
       "r_targets/README.md",
       "README.md",
       strategy = "skip",
-      data = list()
+      data = "auto"
     ),
     step_text(gitignore_lines, ".gitignore", strategy = "union"),
     step_run(fs::dir_create, "R", .label = "fs::dir_create('R')")